% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haon.gg.R
\name{haon.gg}
\alias{haon.gg}
\title{haon.gg}
\usage{
haon.gg()
}
\arguments{
\item{group}{1=adatalab; 2=shiny; 3=visulization}
}
\description{
A function for load the libraries.
}
\examples{
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) + geom_point() + haon.gg()
}
