% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haon.tools.R
\name{haon.tools}
\alias{haon.tools}
\title{haon.tools}
\usage{
haon.tools(group = "adatalab")
}
\arguments{
\item{group}{1=adatalab; 2=shiny; 3=visulization}
}
\description{
A function for load the libraries.
}
\examples{
haon.tools(group = 2)
}
